+++ mapgui-api/src/main/java/de/flog99/mapgui/ServerBackend.java
@@ -1,5 +1,6 @@
 package de.flog99.mapgui;
 
+import de.flog99.mapgui.camera.EntityAngles;
 import de.flog99.mapgui.map.SavedMapPixels;
 import org.jetbrains.annotations.ApiStatus;
 
@@ -28,4 +29,6 @@ public interface ServerBackend {
     RotationController rotation();
 
     SavedMapPixels savedMapPixels();
+
+    EntityAngles entityAngles();
 }